Extract testPush to its own class and rename PushersManager to DefaultPusherSubscriber
