+++ libraries/push/impl/src/main/kotlin/io/element/android/libraries/push/impl/test/TestPush.kt
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2024 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.libraries.push.impl.test
+
+import com.squareup.anvil.annotations.ContributesBinding
+import io.element.android.appconfig.PushConfig
+import io.element.android.libraries.di.AppScope
+import io.element.android.libraries.matrix.api.core.EventId
+import io.element.android.libraries.matrix.api.core.RoomId
+import io.element.android.libraries.push.impl.pushgateway.PushGatewayNotifyRequest
+import io.element.android.libraries.pushproviders.api.CurrentUserPushConfig
+import javax.inject.Inject
+
+interface TestPush {
+    suspend fun execute(config: CurrentUserPushConfig)
+}
+
+@ContributesBinding(AppScope::class)
+class DefaultTestPush @Inject constructor(
+    private val pushGatewayNotifyRequest: PushGatewayNotifyRequest,
+) : TestPush {
+    override suspend fun execute(config: CurrentUserPushConfig) {
+        pushGatewayNotifyRequest.execute(
+            PushGatewayNotifyRequest.Params(
+                url = config.url,
+                appId = PushConfig.PUSHER_APP_ID,
+                pushKey = config.pushKey,
+                eventId = TEST_EVENT_ID,
+                roomId = TEST_ROOM_ID,
+            )
+        )
+    }
+
+    companion object {
+        val TEST_EVENT_ID = EventId("\$THIS_IS_A_FAKE_EVENT_ID")
+        val TEST_ROOM_ID = RoomId("!room:domain")
+    }
+}
